requirements.txt
brotlipy==0.7.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.4
click==7.1.2
colorama==0.4.4
configparser==5.0.2
cryptography==36.0.0
cycler==0.11.0
docker-pycreds==0.4.0
et-xmlfile==1.0.1
fonttools==4.25.0
gitdb==4.0.7
GitPython==3.1.18
idna==3.3
kiwisolver==1.3.2
matplotlib==3.5.1
munkres==1.1.4
numpy==1.20.3
olefile==0.46
openai==0.13.0
openpyxl==3.0.9
packaging==21.3
pandas==1.2.3
pandas-stubs==1.2.0.48
pathtools==0.1.2
Pillow==6.2.1
pip==21.2.4
promise==2.3
protobuf==3.19.1
psutil==5.8.0
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.7
PySocks==1.7.1
python-box==5.4.1
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.6
sentry-sdk==1.5.6
setuptools==58.0.4
shortuuid==1.0.8
sip==4.19.13
six==1.16.0
smmap==4.0.0
subprocess32==3.5.4
termcolor==1.1.0
toml==0.10.2
torch==1.10.2
tornado==6.1
tqdm==4.62.3
typing-extensions==3.10.0.2
urllib3==1.26.8
wandb==0.12.10
wheel==0.37.1
yaspin==2.1.0
